+++ src/main/java/edu/brandeis/cs/nlp/mae/controller/deprecated/HelpMenuListener.java
@@ -22,45 +22,44 @@
  * @see <a href="https://github.com/keighrim/mae-annotation">https://github.com/keighrim/mae-annotation</a>
  */
 
-package edu.brandeis.cs.nlp.mae.controller.menu.mode;
+package edu.brandeis.cs.nlp.mae.controller.deprecated;
 
+import edu.brandeis.cs.nlp.mae.MaeStrings;
 import edu.brandeis.cs.nlp.mae.controller.MaeMainUI;
 
+import java.awt.*;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
+import java.io.IOException;
+import java.net.URI;
+import java.net.URISyntaxException;
 
 /**
- * Listener to select special modes
+ * Listens for the request from the Help Menu
  */
-public class ModeMenuListener implements ActionListener {
+public class HelpMenuListener implements ActionListener {
     private MaeMainUI maeMainUI;
 
-    public ModeMenuListener(MaeMainUI maeMainUI) {
+    public HelpMenuListener(MaeMainUI maeMainUI) {
         this.maeMainUI = maeMainUI;
     }
 
-    // TODO add adjud mode
+    @Override
     public void actionPerformed(ActionEvent actionEvent) {
-        int action = Integer.parseInt(actionEvent.getActionCommand());
-
-        switch (action) {
-            // return to normal mode
-            case MaeMainUI.M_NORMAL:
-                maeMainUI.returnToNormalMode(true);
-                break;
-            case MaeMainUI.M_MULTI_SPAN:
-                maeMainUI.setMode(MaeMainUI.M_MULTI_SPAN);
-                maeMainUI.getStatusBar().setText(
-                        "Multi-span mode! Click anywhere to continue.");
-                break;
-            case MaeMainUI.M_ARG_SEL:
-                maeMainUI.setMode(MaeMainUI.M_ARG_SEL);
-                maeMainUI.getStatusBar().setText(
-                        "Argument select mode! Click anywhere to continue.");
-                break;
+        String command = actionEvent.getActionCommand();
+        if (command.equals("about")) {
+            maeMainUI.showAboutDialog();
+        } else if (command.equals("web")) {
+            if (Desktop.isDesktopSupported()) {
+                try {
+                    Desktop.getDesktop().browse(
+                            new URI(MaeStrings.PROJECT_WEBPAGE));
+                } catch (IOException e) {
+                    e.printStackTrace();
+                } catch (URISyntaxException e) {
+                    e.printStackTrace();
+                }
+            }
         }
-        maeMainUI.updateMenus();
-        maeMainUI.resetSpans();
-        maeMainUI.delayedUpdateStatusBar(3000);
     }
 }
