+++ src/main/java/edu/brandeis/cs/nlp/mae/controller/deprecated/SetAsArgListener.java
@@ -22,40 +22,33 @@
  * @see <a href="https://github.com/keighrim/mae-annotation">https://github.com/keighrim/mae-annotation</a>
  */
 
-package edu.brandeis.cs.nlp.mae.controller;
+package edu.brandeis.cs.nlp.mae.controller.deprecated;
 
+import edu.brandeis.cs.nlp.mae.MaeStrings;
 import edu.brandeis.cs.nlp.mae.controller.MaeMainUI;
 
-import java.awt.event.MouseAdapter;
-import java.awt.event.MouseEvent;
+import java.awt.event.ActionEvent;
+import java.awt.event.ActionListener;
 
 /**
- * PopupListener determines whether the link creation window should be
- * displayed.
+ * This listener is associated with 'set as argument of...' menu items
+ * callable from context menus from either bottom table or main text pane
  */
-public class TextMouseAdapter extends MouseAdapter {
-
+public class SetAsArgListener implements ActionListener {
     private MaeMainUI maeMainUI;
 
-    public TextMouseAdapter(MaeMainUI maeMainUI) {
+    public SetAsArgListener(MaeMainUI maeMainUI) {
         this.maeMainUI = maeMainUI;
     }
 
     @Override
-    public void mousePressed(MouseEvent e) {
-        maybeShowTextPopup(e);
-    }
-
-    @Override
-    public void mouseReleased(MouseEvent e) {
-        maybeShowTextPopup(e);
-    }
-
-    private void maybeShowTextPopup(MouseEvent e) {
-        if (e.isPopupTrigger()) {
-            maeMainUI.setTextPopup(maeMainUI.createTextContextMenu());
-            maeMainUI.getTextPopup().show(e.getComponent(),
-                    e.getX(), e.getY());
-        }
+    public void actionPerformed(ActionEvent actionEvent) {
+        // command looks like this:
+        // linkType(0), linkId(1), argName(2), argId(3), argText(4)
+        String[] command = actionEvent.getActionCommand().split(MaeStrings.SEP);
+        maeMainUI.setArgumentInTable(command[0], command[1], command[2], command[3], command[4]);
+        int argNum = maeMainUI.getTask().getArgumentTypesOfLinkTagType(command[0]).indexOf(command[2]);
+        String argType = maeMainUI.getTask().getTagTypeByTid(command[3]);
+        maeMainUI.getTask().addArgument(command[1], argNum, command[3], argType);
     }
 }
